payload_processor: adopt global operations from DeltaPerformer
PayloadProcessor is now responsible for maintaining global update state
and any manifest level data. DeltaPerformer is left with just
dispatching a sequence of update operations given to to it on the device
it has opened.
